develop : adding int parse to ftp_timeout

+++ config/filesystems.php
@@ -61,7 +61,7 @@ return [
             'port' => (int) env('FTP_PORT', 21),
             'passive' => false,
             'ignorePassiveAddress' => true, 
-            'timeout' => env('FTP_TIMEOUT', 360),
+            'timeout' => (int) env('FTP_TIMEOUT', 360),
             'ssl'      => env('FTP_SSL', false),
         ]
